test: extract test_matrix.py
Unit tests for `cr.cube.matrix` remain in `test_cubepart.py` after
extracting the separate matrix module. Move those tests to their own
unit test module.
